Download lots of data from an Intercom account
This script doesn't download everything that can be retrieved via the
API [1], but it does list the:

- admins
- companies
- company data attributes
- conversations
- customer data attributes
- leads
- segments
- tags
- teams
- users

It doesn't retrieve the notes or data events, as they both need to be
retrieved per-user.

I'm not too bothered about that data as I write this script, so I
haven't done the work to extend the script to retrieve it.

[1] https://developers.intercom.com/intercom-api-reference/v1.1/reference